Refactor and improve code quality across multiple modules
This commit significantly improves code organization, maintainability, and quality:

1. Split behavior_algorithms.py into modular structure (2726 → 89 lines)
   - Created core/algorithms/ directory with 7 modules
   - Organized 48 algorithms into categories: food_seeking, predator_avoidance,
     schooling, energy_management, territory, and base
   - Maintained full backward compatibility via re-exports
   - Each category file is now focused and maintainable (200-600 lines)

2. Improved core/environment.py
   - Consolidated duplicate methods agents_to_avoid and agents_to_align_with
   - Created new nearby_agents_by_type method with backward-compatible aliases
   - Fixed broad exception handling (Exception → ImportError)

3. Enhanced fishtank_pz.py
   - Added missing docstrings to __init__, render, and close methods
   - Fixed broad exception handling (Exception → ImportError)

4. Refactored movement_strategy.py
   - Extracted magic numbers to named constants
   - Added NEURAL_MOVEMENT_SMOOTHING, ALGORITHMIC_MOVEMENT_SMOOTHING,
     and ALGORITHMIC_MAX_SPEED_MULTIPLIER constants
   - Added explanatory comments for FOOD_ALIGNMENT_DISTANCE

5. Code quality improvements
   - Removed unused imports
   - All files compile successfully
   - Better code organization and maintainability

Impact: Major improvement in code maintainability while preserving all functionality
and backward compatibility. The 2700-line monolithic file is now a clean,
category-based module structure.
