+++ core/algorithms/territory.py
@@ -1,0 +1,427 @@
+"""Territory and exploration behavior algorithms.
+
+This module contains 8 algorithms focused on spatial behavior and exploration:
+- TerritorialDefender: Defend a territory from other fish
+- RandomExplorer: Explore randomly, covering new ground
+- WallFollower: Follow along tank walls
+- CornerSeeker: Prefer staying in corners
+- CenterHugger: Stay near the center of the tank
+- RoutePatroller: Patrol between specific waypoints
+- BoundaryExplorer: Explore edges and boundaries
+- NomadicWanderer: Wander continuously without a home base
+"""
+
+import random
+import math
+from typing import Tuple, List
+from dataclasses import dataclass
+
+from core.algorithms.base import BehaviorAlgorithm, Vector2
+from core.constants import SCREEN_WIDTH, SCREEN_HEIGHT
+from agents import Food, Crab, Fish as FishClass
+
+
+@dataclass
+class TerritorialDefender(BehaviorAlgorithm):
+    """Defend a territory from other fish."""
+
+    def __init__(self):
+        super().__init__(
+            algorithm_id="territorial_defender",
+            parameters={
+                "territory_radius": random.uniform(80, 150),
+                "aggression": random.uniform(0.5, 1.0),
+            }
+        )
+        self.territory_center = None
+
+    @classmethod
+    def random_instance(cls):
+        return cls()
+
+    def execute(self, fish: 'Fish') -> Tuple[float, float]:
+
+        if self.territory_center is None:
+            self.territory_center = fish.pos.copy()
+
+        # Chase away intruders
+        intruders = [f for f in fish.environment.get_agents_of_type(FishClass)
+                    if f != fish and (f.pos - self.territory_center).length() < self.parameters["territory_radius"]]
+
+        if intruders:
+            nearest = min(intruders, key=lambda f: (f.pos - fish.pos).length())
+            direction = (nearest.pos - fish.pos).normalize()
+            return direction.x * self.parameters["aggression"], direction.y * self.parameters["aggression"]
+
+        # Return to territory center
+        direction = (self.territory_center - fish.pos)
+        if direction.length() > self.parameters["territory_radius"]:
+            direction = direction.normalize()
+            return direction.x * 0.5, direction.y * 0.5
+
+        return 0, 0
+
+
+@dataclass
+class RandomExplorer(BehaviorAlgorithm):
+    """Explore randomly, covering new ground."""
+
+    def __init__(self):
+        super().__init__(
+            algorithm_id="random_explorer",
+            parameters={
+                "change_frequency": random.uniform(0.02, 0.08),
+                "exploration_speed": random.uniform(0.5, 0.9),
+            }
+        )
+        self.current_direction = Vector2(random.uniform(-1, 1), random.uniform(-1, 1)).normalize()
+
+    @classmethod
+    def random_instance(cls):
+        return cls()
+
+    def execute(self, fish: 'Fish') -> Tuple[float, float]:
+
+        # Check for important stimuli
+        nearest_predator = self._find_nearest(fish, Crab)
+        nearest_food = self._find_nearest(fish, Food)
+
+        # Predator avoidance
+        if nearest_predator and (nearest_predator.pos - fish.pos).length() < 120:
+            direction = (fish.pos - nearest_predator.pos).normalize()
+            return direction.x * 1.2, direction.y * 1.2
+
+        # Food opportunism
+        if nearest_food:
+            food_distance = (nearest_food.pos - fish.pos).length()
+            if food_distance < 70:
+                direction = (nearest_food.pos - fish.pos).normalize()
+                return direction.x * 1.0, direction.y * 1.0
+
+        # Boundary avoidance - don't explore into walls
+        edge_margin = 60
+        avoid_x, avoid_y = 0, 0
+        if fish.pos.x < edge_margin:
+            avoid_x = 0.5
+        elif fish.pos.x > SCREEN_WIDTH - edge_margin:
+            avoid_x = -0.5
+        if fish.pos.y < edge_margin:
+            avoid_y = 0.5
+        elif fish.pos.y > SCREEN_HEIGHT - edge_margin:
+            avoid_y = -0.5
+
+        # Change direction periodically or when hitting boundaries
+        if random.random() < self.parameters["change_frequency"] or (avoid_x != 0 or avoid_y != 0):
+            # Bias new direction away from edges
+            new_x = random.uniform(-1, 1) + avoid_x
+            new_y = random.uniform(-1, 1) + avoid_y
+            self.current_direction = Vector2(new_x, new_y)
+            if self.current_direction.length() > 0:
+                self.current_direction = self.current_direction.normalize()
+
+        # Sometimes explore toward unexplored areas (away from other fish)
+        if random.random() < 0.1:
+            allies = fish.environment.get_agents_of_type(FishClass)
+            if len(allies) > 1:
+                # Find average position of other fish
+                other_fish = [f for f in allies if f != fish]
+                if other_fish:
+                    crowd_center = sum((f.pos for f in other_fish), Vector2()) / len(other_fish)
+                    # Explore away from the crowd
+                    away_from_crowd = (fish.pos - crowd_center)
+                    if away_from_crowd.length() > 0:
+                        self.current_direction = away_from_crowd.normalize()
+
+        return self.current_direction.x * self.parameters["exploration_speed"], \
+               self.current_direction.y * self.parameters["exploration_speed"]
+
+
+@dataclass
+class WallFollower(BehaviorAlgorithm):
+    """Follow along tank walls."""
+
+    def __init__(self):
+        super().__init__(
+            algorithm_id="wall_follower",
+            parameters={
+                "wall_distance": random.uniform(20, 60),
+                "follow_speed": random.uniform(0.5, 0.8),
+            }
+        )
+
+    @classmethod
+    def random_instance(cls):
+        return cls()
+
+    def execute(self, fish: 'Fish') -> Tuple[float, float]:
+
+        # Find nearest wall
+        dist_to_left = fish.pos.x
+        dist_to_right = SCREEN_WIDTH - fish.pos.x
+        dist_to_top = fish.pos.y
+        dist_to_bottom = SCREEN_HEIGHT - fish.pos.y
+
+        min_dist = min(dist_to_left, dist_to_right, dist_to_top, dist_to_bottom)
+
+        # Move parallel to nearest wall
+        if min_dist == dist_to_left or min_dist == dist_to_right:
+            return 0, self.parameters["follow_speed"]
+        else:
+            return self.parameters["follow_speed"], 0
+
+
+@dataclass
+class CornerSeeker(BehaviorAlgorithm):
+    """Prefer staying in corners."""
+
+    def __init__(self):
+        super().__init__(
+            algorithm_id="corner_seeker",
+            parameters={
+                "preferred_corner": random.choice(['top_left', 'top_right', 'bottom_left', 'bottom_right']),
+                "approach_speed": random.uniform(0.4, 0.7),
+            }
+        )
+
+    @classmethod
+    def random_instance(cls):
+        return cls()
+
+    def execute(self, fish: 'Fish') -> Tuple[float, float]:
+
+        # Determine corner position
+        corners = {
+            'top_left': Vector2(50, 50),
+            'top_right': Vector2(SCREEN_WIDTH - 50, 50),
+            'bottom_left': Vector2(50, SCREEN_HEIGHT - 50),
+            'bottom_right': Vector2(SCREEN_WIDTH - 50, SCREEN_HEIGHT - 50),
+        }
+
+        target = corners[self.parameters["preferred_corner"]]
+        direction = (target - fish.pos)
+        if direction.length() > 0:
+            direction = direction.normalize()
+            return direction.x * self.parameters["approach_speed"], direction.y * self.parameters["approach_speed"]
+        return 0, 0
+
+
+@dataclass
+class CenterHugger(BehaviorAlgorithm):
+    """Stay near the center of the tank."""
+
+    def __init__(self):
+        super().__init__(
+            algorithm_id="center_hugger",
+            parameters={
+                "orbit_radius": random.uniform(50, 120),
+                "return_strength": random.uniform(0.5, 0.9),
+            }
+        )
+
+    @classmethod
+    def random_instance(cls):
+        return cls()
+
+    def execute(self, fish: 'Fish') -> Tuple[float, float]:
+
+        center = Vector2(SCREEN_WIDTH / 2, SCREEN_HEIGHT / 2)
+        distance = (center - fish.pos).length()
+
+        if distance > self.parameters["orbit_radius"]:
+            direction = (center - fish.pos).normalize()
+            return direction.x * self.parameters["return_strength"], direction.y * self.parameters["return_strength"]
+        return 0, 0
+
+
+@dataclass
+class RoutePatroller(BehaviorAlgorithm):
+    """Patrol between specific waypoints."""
+
+    def __init__(self):
+        super().__init__(
+            algorithm_id="route_patroller",
+            parameters={
+                "patrol_speed": random.uniform(0.5, 0.8),
+                "waypoint_threshold": random.uniform(30, 60),
+            }
+        )
+        self.waypoints: List[Vector2] = []
+        self.current_waypoint_idx = 0
+        self.initialized = False
+
+    @classmethod
+    def random_instance(cls):
+        return cls()
+
+    def execute(self, fish: 'Fish') -> Tuple[float, float]:
+
+        if not self.initialized:
+            # Create strategic waypoints - cover different areas of tank
+            num_waypoints = random.randint(4, 7)
+            for i in range(num_waypoints):
+                # Distribute waypoints in a pattern
+                angle = (2 * math.pi * i) / num_waypoints
+                radius = random.uniform(120, 250)
+                center_x, center_y = SCREEN_WIDTH / 2, SCREEN_HEIGHT / 2
+                wp_x = center_x + math.cos(angle) * radius
+                wp_y = center_y + math.sin(angle) * radius
+                # Clamp to bounds
+                wp_x = max(60, min(SCREEN_WIDTH - 60, wp_x))
+                wp_y = max(60, min(SCREEN_HEIGHT - 60, wp_y))
+                self.waypoints.append(Vector2(wp_x, wp_y))
+            self.initialized = True
+
+        if not self.waypoints:
+            return 0, 0
+
+        # Check for threats and opportunities
+        nearest_predator = self._find_nearest(fish, Crab)
+        nearest_food = self._find_nearest(fish, Food)
+
+        # Interrupt patrol for immediate threats
+        if nearest_predator and (nearest_predator.pos - fish.pos).length() < 100:
+            direction = (fish.pos - nearest_predator.pos).normalize()
+            return direction.x * 1.3, direction.y * 1.3
+
+        # Interrupt patrol for close food
+        if nearest_food and (nearest_food.pos - fish.pos).length() < 60:
+            direction = (nearest_food.pos - fish.pos).normalize()
+            return direction.x * 0.9, direction.y * 0.9
+
+        # Continue patrol
+        target = self.waypoints[self.current_waypoint_idx]
+        distance = (target - fish.pos).length()
+
+        # Reached waypoint
+        if distance < self.parameters["waypoint_threshold"]:
+            # Look for food near this waypoint before moving on
+            if nearest_food and (nearest_food.pos - target).length() < 100:
+                # Food near waypoint - pursue it
+                direction = (nearest_food.pos - fish.pos).normalize()
+                return direction.x * self.parameters["patrol_speed"], direction.y * self.parameters["patrol_speed"]
+
+            # Move to next waypoint
+            self.current_waypoint_idx = (self.current_waypoint_idx + 1) % len(self.waypoints)
+            target = self.waypoints[self.current_waypoint_idx]
+
+        # Move toward current waypoint
+        direction = (target - fish.pos)
+        if direction.length() > 0:
+            direction = direction.normalize()
+
+            # Vary speed based on distance - slow down when approaching
+            speed_multiplier = 1.0
+            if distance < 80:
+                speed_multiplier = 0.7 + (distance / 80) * 0.3
+
+            return direction.x * self.parameters["patrol_speed"] * speed_multiplier, \
+                   direction.y * self.parameters["patrol_speed"] * speed_multiplier
+        return 0, 0
+
+
+@dataclass
+class BoundaryExplorer(BehaviorAlgorithm):
+    """Explore edges and boundaries."""
+
+    def __init__(self):
+        super().__init__(
+            algorithm_id="boundary_explorer",
+            parameters={
+                "edge_preference": random.uniform(0.6, 1.0),
+                "exploration_speed": random.uniform(0.5, 0.8),
+            }
+        )
+
+    @classmethod
+    def random_instance(cls):
+        return cls()
+
+    def execute(self, fish: 'Fish') -> Tuple[float, float]:
+
+        # Move toward edges
+        center = Vector2(SCREEN_WIDTH / 2, SCREEN_HEIGHT / 2)
+        direction = (fish.pos - center)
+        if direction.length() > 0:
+            direction = direction.normalize()
+            return direction.x * self.parameters["exploration_speed"] * self.parameters["edge_preference"], \
+                   direction.y * self.parameters["exploration_speed"] * self.parameters["edge_preference"]
+        return 0, 0
+
+
+@dataclass
+class NomadicWanderer(BehaviorAlgorithm):
+    """Wander continuously without a home base."""
+
+    def __init__(self):
+        super().__init__(
+            algorithm_id="nomadic_wanderer",
+            parameters={
+                "wander_strength": random.uniform(0.5, 0.9),
+                "direction_change_rate": random.uniform(0.01, 0.05),
+            }
+        )
+        self.wander_angle = random.uniform(0, 2 * math.pi)
+
+    @classmethod
+    def random_instance(cls):
+        return cls()
+
+    def execute(self, fish: 'Fish') -> Tuple[float, float]:
+
+        # Check for threats and opportunities
+        nearest_predator = self._find_nearest(fish, Crab)
+        nearest_food = self._find_nearest(fish, Food)
+
+        # Immediate threat response
+        if nearest_predator and (nearest_predator.pos - fish.pos).length() < 110:
+            # Flee but maintain nomadic unpredictability
+            base_escape = (fish.pos - nearest_predator.pos).normalize()
+            perp_x, perp_y = -base_escape.y, base_escape.x
+            randomness = random.uniform(-0.4, 0.4)
+            vx = base_escape.x * 1.0 + perp_x * randomness
+            vy = base_escape.y * 1.0 + perp_y * randomness
+            return vx, vy
+
+        # Opportunistic food grab
+        if nearest_food:
+            food_distance = (nearest_food.pos - fish.pos).length()
+            if food_distance < 50:
+                # Close food - grab it
+                direction = (nearest_food.pos - fish.pos).normalize()
+                return direction.x * 1.0, direction.y * 1.0
+
+        # Boundary awareness - avoid getting stuck in corners
+        edge_margin = 70
+        boundary_influence = 0
+        if fish.pos.x < edge_margin or fish.pos.x > SCREEN_WIDTH - edge_margin or \
+           fish.pos.y < edge_margin or fish.pos.y > SCREEN_HEIGHT - edge_margin:
+            # Turn toward center
+            center = Vector2(SCREEN_WIDTH / 2, SCREEN_HEIGHT / 2)
+            to_center = (center - fish.pos).normalize()
+            # Blend turn toward center with random wandering
+            boundary_influence = 0.4
+            self.wander_angle = math.atan2(to_center.y, to_center.x) + random.gauss(0, 0.3)
+
+        # Gradually change direction with smooth random walk
+        angle_change = random.gauss(0, self.parameters["direction_change_rate"])
+        self.wander_angle += angle_change
+
+        # Add some Perlin-like noise for more natural wandering
+        # Use fish position to add spatial variation
+        spatial_influence = math.sin(fish.pos.x * 0.01) * 0.05 + math.cos(fish.pos.y * 0.01) * 0.05
+        self.wander_angle += spatial_influence
+
+        vx = math.cos(self.wander_angle) * self.parameters["wander_strength"]
+        vy = math.sin(self.wander_angle) * self.parameters["wander_strength"]
+
+        # Energy-based activity
+        energy_ratio = fish.energy / fish.max_energy
+        if energy_ratio < 0.4:
+            # Lower energy = more purposeful toward food
+            if nearest_food and (nearest_food.pos - fish.pos).length() < 150:
+                food_dir = (nearest_food.pos - fish.pos).normalize()
+                # Blend wandering with food seeking
+                vx = vx * 0.4 + food_dir.x * 0.6
+                vy = vy * 0.4 + food_dir.y * 0.6
+
+        return vx, vy
